Multiple changes:
- make mod update checking concurrent
- fix `ModrinthBackend::get_latest_version_date` not respecting loader
(leading to garbage mod updates)
- add "Checking for mod updates" message to Mods menu
- add pulsating dots to a few Loading... texts (using Launcher.tick_timer)
- fix clippy too many args
- change some i64 to isize (for platform compat)

+++ crates/ql_instances/src/mc_auth/mod.rs
@@ -66,7 +66,7 @@ pub struct AuthCodeResponse {
     pub user_code: String,
     pub device_code: String,
     pub verification_uri: String,
-    pub expires_in: i64,
+    pub expires_in: isize,
     pub interval: u64,
     pub message: String,
 }
